[REFACTOR] 소켓 스페이스 구조 분리
