+++ src/middlewares/room-token.middleware.ts
@@ -2,7 +2,7 @@ import { Request, Response, NextFunction } from "express";
 import ChatRoomModel from "../models/chat/chat-room.model";
 import responseHandler from "../handlers/response.handler";
 import userModel from "../models/chat/chat-room.model";
-import { CustomRequest } from "../types/request.type";
+import { CustomRequest } from "../@types/request.type";
 import ChatRoomDto from "../dto/chat/room.dto";
 
 const findRoom = async (req: Request) => {
